split lib.rs into multiple module
